Refactor: Move expression tests to separate module (Phase 1.1)
Implements Issue #23 - Phase 1.1 of test file refactoring.

Changes:
- Created crates/executor/src/tests/ directory structure
- Moved 6 expression evaluator tests to tests/expression.rs
- Updated module organization with tests/mod.rs
- Remaining SELECT and pagination tests kept in select_and_pagination.rs

Test Coverage:
- 6 expression tests moved (test_eval_integer_literal, test_eval_string_literal,
  test_eval_null_literal, test_eval_column_ref, test_eval_column_not_found,
  test_eval_addition)
- All tests maintain same behavior
- Clippy passes with no warnings

File Structure:
crates/executor/src/tests/
├── mod.rs (module declarations)
├── expression.rs (111 lines, 6 tests)
└── select_and_pagination.rs (remaining tests for future phases)

Note: Original issue stated 11 tests, but lines 5-101 contained 6 tests.
Curator's discrepancy note was correct.

Closes #23
