+++ crates/executor/src/tests/mod.rs
@@ -1,0 +1,6 @@
+//! Test modules for executor crate
+//!
+//! Tests are organized by feature area for better maintainability.
+
+pub mod expression;
+pub mod select_and_pagination;
